implement and testing new busnis logic of user msc, and incorpore SSL selfsigning in api gateway

+++ Service/User/Module/Service/UserServices.cs
@@ -1,7 +1,5 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc.ModelBinding.Binders;
-using Microsoft.OpenApi.Expressions;
 using User.Module.DTOs;
 using User.Module.Model;
 using User.Module.Repository.Interface;
@@ -87,6 +85,9 @@ namespace User.Module.Service
             var user = await this._repository.FindByIdAsync(id) ??
                 throw new NotFoundExceptions("User not found");
 
+            if (user.Roles == ROLES.Admin)
+                throw new ForbiddenExceptions("Could not delete a user Admin");
+
             await this._repository.DeleteAsync(user);
         }
         /// <summary>
@@ -121,7 +122,7 @@ namespace User.Module.Service
 
             var verificationPass = passwordHasher.VerifyHashedPassword(date, date.Password, oldPassword);
             if (verificationPass == PasswordVerificationResult.Failed)
-                throw new BadRequestExceptions("Password Wrong");
+                throw new ForbiddenExceptions("Password Wrong");
 
             this._validation.ValidateStructurePassword(password);
 
@@ -144,14 +145,93 @@ namespace User.Module.Service
         /// <exception cref="NotImplementedException"></exception>
         public async Task<UserModel> UpdateRegister(UpdateUserDTO body, int id)
         {
-            var user = await this._repository.FindByIdAsync(id) ?? 
+            var user = await this._repository.FindByIdAsync(id) ??
                 throw new NotFoundExceptions("User not found");
 
+            if (user.Roles == ROLES.Admin)
+                throw new ForbiddenExceptions("Could not edit a user Admin");
+
             this._validation.ValidationEmail(body.Email);
-            
+
             this._mapper.Map(body, user);
             await this._repository.UpdateAsync(user);
             return user;
+        }
+        /// <summary>
+        /// Update own register
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="password"></param>
+        /// <param name="body"></param>
+        /// <returns></returns>
+        /// <exception cref="NotFoundExceptions"></exception>
+        /// <exception cref="ForbiddenExceptions"></exception>
+        public async Task<string> UpdateOwnRegister(int id, string password, UpdateUserDTO body)
+        {
+            if (password == null)
+                throw new BadRequestExceptions("Password is required");
+
+            var user = await this._repository.FindByIdAsync(id) ??
+                throw new NotFoundExceptions("User not found");
+
+            var passwordHasher = new PasswordHasher<UserModel>();
+            var verificationPass = passwordHasher.VerifyHashedPassword(user, user.Password, password);
+            if (verificationPass == PasswordVerificationResult.Failed)
+                throw new ForbiddenExceptions("Password is wrong");
+
+            this._validation.ValidationEmail(body.Email);
+
+            this._mapper.Map(body, user);
+            await this._repository.UpdateAsync(user);
+
+            return "Your reforms was saved successfully";
+        }
+        /// <summary>
+        /// Update Roles
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="newRoles"></param>
+        /// <returns></returns>
+        /// <exception cref="NotFoundExceptions"></exception>
+        public async Task<string> UpdateRoles(int id, RolesDTO newRoles)
+        {
+            var user = await this._repository.FindByIdAsync(id) ??
+                throw new NotFoundExceptions("User not found");
+
+            if (user.Roles == newRoles.Roles)
+                throw new BadRequestExceptions("No changes were made, Roles is already set");
+
+            user.Roles = newRoles.Roles;
+
+            await this._repository.UpdateAsync(user);
+
+            return "Roles were updated successfully";
+        }
+
+        /// <summary>
+        /// Remove user for basic user roles
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="password"></param>
+        /// <returns></returns>
+        /// <exception cref="NotImplementedException"></exception>
+        public async Task<string> RemoveUserRegisterForBasicRoles(int id, string password)
+        {
+            if (password == null)
+                throw new BadRequestExceptions("Password is required");
+                
+            var user = await this._repository.FindByIdAsync(id) ??
+                throw new NotFoundExceptions("User not found");
+
+            var passwordHasher = new PasswordHasher<UserModel>();
+
+            var verificationPass = passwordHasher.VerifyHashedPassword(user, user.Password, password);
+            if (verificationPass == PasswordVerificationResult.Failed)
+                throw new ForbiddenExceptions("Password Wrong");
+
+            await this._repository.DeleteAsync(user);
+
+            return "User was remove successfully";
         }
     }
 }
